+++ restli-server/src/main/java/com/linkedin/restli/internal/server/response/PartialUpdateResponseBuilder.java
@@ -87,7 +87,7 @@ public class PartialUpdateResponseBuilder implements RestLiResponseBuilder<RestL
         TimingContextUtil.beginTiming(resourceContext.getRawRequestContext(),
             FrameworkTimingKeys.SERVER_RESPONSE_RESTLI_PROJECTION_APPLY.key());
 
-        final DataMap data = RestUtils.projectFields(entityData, resourceContext.getProjectionMode(), resourceContext.getProjectionMask());
+        final DataMap data = RestUtils.projectFields(entityData, resourceContext);
 
         TimingContextUtil.endTiming(resourceContext.getRawRequestContext(),
             FrameworkTimingKeys.SERVER_RESPONSE_RESTLI_PROJECTION_APPLY.key());