+++ restli-server/src/main/java/com/linkedin/restli/server/config/RestLiMethodConfigBuilder.java
@@ -3,6 +3,7 @@ package com.linkedin.restli.server.config;
 import java.util.HashMap;
 import java.util.Map;
 
+
 /**
  * A builder to build {@link RestLiMethodConfig}. This is created for extensibility.
  *
@@ -11,7 +12,7 @@ import java.util.Map;
 public class RestLiMethodConfigBuilder
 {
   private final Map<String, Long> _timeoutMsConfig = new HashMap<>();
-
+  private final Map<String, String> _alwaysProjectedFieldsConfig = new HashMap<>();
   // Whether to validate parameter in the query parameters.
   private boolean shouldValidateQueryParams = false;
   private boolean shouldValidateResourceKeys = false;
@@ -32,12 +33,14 @@ public class RestLiMethodConfigBuilder
       addTimeoutMsConfigMap(config.getTimeoutMsConfig());
       withShouldValidateQueryParams(config.shouldValidateQueryParams());
       withShouldValidateResourceKeys(config.shouldValidateResourceKey());
+      addAlwaysProjectedFieldsMap(config.getAlwaysProjectedFieldsConfig());
     }
   }
 
   public RestLiMethodConfig build()
   {
-    return new RestLiMethodConfigImpl(_timeoutMsConfig, shouldValidateQueryParams, shouldValidateResourceKeys);
+    return new RestLiMethodConfigImpl(_timeoutMsConfig, shouldValidateQueryParams, shouldValidateResourceKeys,
+        _alwaysProjectedFieldsConfig);
   }
 
   public RestLiMethodConfigBuilder withShouldValidateQueryParams(boolean shouldValidateQueryParams)
@@ -67,6 +70,24 @@ public class RestLiMethodConfigBuilder
   public RestLiMethodConfigBuilder clearTimeoutMs()
   {
     _timeoutMsConfig.clear();
+    return this;
+  }
+
+  public RestLiMethodConfigBuilder addAlwaysProjectedFieldsMap(Map<String, String> config)
+  {
+    _alwaysProjectedFieldsConfig.putAll(config);
+    return this;
+  }
+
+  public RestLiMethodConfigBuilder addAlwaysProjectedFields(String key, String value)
+  {
+    _alwaysProjectedFieldsConfig.put(key, value);
+    return this;
+  }
+
+  public RestLiMethodConfigBuilder clearAlwaysProjectedFields()
+  {
+    _alwaysProjectedFieldsConfig.clear();
     return this;
   }
 }